lock on target after control input
Why we want this: If mouse/controller input is accepted while we are locked to a target we get jittery motion as one frame will have the motion and the next its snapped back to the target. This ensures we just stay snapped to the target. I'm assuming this is the intended behavior.

Really we could ignore input if we're locked, but i found this method to be more reliable AND if we want some limited movement while staying locked this easily allows that implemented. ( I have that version - happy to share if thats desired but felt like keeping the template simple was best.)